Re-work value generator section
* Move generator code to the dedicated class -> Generator
* Expose formula generator property to the API

+++ src/main/java/org/obd/metrics/statistics/DefaultStatistics.java
@@ -15,8 +15,8 @@ final class DefaultStatistics implements Statistics {
 	}
 
 	@Override
-	public double getMedian() {
-		return delegate.getMedian();
+	public long getMedian() {
+		return (long) delegate.getMedian();
 	}
 
 	@Override
