+++ src/main/java/org/obd/metrics/statistics/Statistics.java
@@ -3,7 +3,7 @@ package org.obd.metrics.statistics;
 public interface Statistics {
 	int size();
 
-	double getMedian();
+	long getMedian();
 
 	long getMax();
